requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.45
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.0
openai==1.86.0
orjson==3.10.18
packaging==24.2
pip==23.2.1
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==65.5.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
tweepy==4.15.0
types-requests==2.32.0.20250602
typing_extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.23.0